+++ MakeoProject/Views/Freelances.xaml.cs
@@ -10,15 +10,9 @@ using System.Windows.Data;
 
 namespace MakeoProject.Views
 {
-    /// <summary>
-    /// Interaction logic for Freelances.xaml
-    /// </summary>
     public partial class Freelances : UserControl
     {
-        // ObservableCollection to store all freelances
         public ObservableCollection<Freelance> allfreelance { get; set; }
-
-        // Property to store the selected freelance
         public Freelance? SelectedFreelance { get; set; }
 
         private void LoadFreelances()
@@ -33,21 +27,13 @@ namespace MakeoProject.Views
             }
         }
 
-        // Constructor
         public Freelances()
         {
             InitializeComponent();
-
-            // Set DataContext to this instance
             this.DataContext = this;
 
-            // Fetch all freelances from the database using Entity Framework Core
-            using (MakeoProjectContext allfreelances = new())
-            {
-                allfreelance = new ObservableCollection<Freelance>(allfreelances.Freelances.Include(p => p.FreelanceCompetences).ThenInclude(pc => pc.IdCompetencesNavigation).ToList());
-            }
+            LoadFreelances();
 
-            // Define columns to display in the DataGrid
             var columnsToDisplay = new Dictionary<string, string>
             {
                 { "Id", "ID" },
@@ -57,7 +43,6 @@ namespace MakeoProject.Views
                 { "Competence", "Compétences" },
             };
 
-            // Loop through columns to create DataGrid columns
             foreach (var column in columnsToDisplay)
             {
                 DataGridTextColumn dataColumn = new DataGridTextColumn
@@ -68,70 +53,83 @@ namespace MakeoProject.Views
                 switch (column.Key)
                 {
                     case "Id":
-                        dataColumn.Width = new DataGridLength(50); // Set width for ID column
+                        dataColumn.Width = new DataGridLength(50);
                         break;
                     case "Name":
                     case "Surname":
-                        dataColumn.Width = new DataGridLength(1, DataGridLengthUnitType.Auto); // Set width to Auto for Name, Surname, and Description columns
-                        dataColumn.MinWidth = 150; // Set minimum width to 150 pixels
+                        dataColumn.Width = new DataGridLength(1, DataGridLengthUnitType.Auto);
+                        dataColumn.MinWidth = 150;
                         break;
                     case "Description":
-                        dataColumn.Width = new DataGridLength(200); // Set width for ID column
+                        dataColumn.Width = new DataGridLength(200);
                         break;
                     default:
-                        dataColumn.Width = new DataGridLength(1, DataGridLengthUnitType.Star); // Set width to Star for remaining columns
-                        dataColumn.Binding = new Binding("Skills"); // Bind Skills property for Competence column
+                        dataColumn.Width = new DataGridLength(1, DataGridLengthUnitType.Star);
+                        dataColumn.Binding = new Binding("Skills");
                         break;
                 }
-                dgFreelances.Columns.Add(dataColumn); // Add the created column to DataGrid
+                dgFreelances.Columns.Add(dataColumn);
             }
 
-            dgFreelances.AutoGenerateColumns = false; // Disable auto generation of columns
+            dgFreelances.AutoGenerateColumns = false;
+
+            // Masquer le bouton "Ajouter" et "Supprimer" si l'utilisateur actuel n'est pas administrateur
+            if (Session.CurrentUser != null && !Session.CurrentUser.Admin)
+            {
+                AddButton.Visibility = Visibility.Collapsed;
+                RemoveButton.Visibility = Visibility.Collapsed;
+            }
         }
 
-        // Ajoutez une méthode pour recharger les données et actualiser le DataGrid
         private void RefreshDataGrid()
         {
-            // Rechargez les freelances depuis la base de données
             LoadFreelances();
-
-            // Actualisez le DataGrid en redéfinissant sa source de données
             dgFreelances.ItemsSource = allfreelance;
         }
 
-        // Event handler for Add Button click
         private void AddButton_Click(object sender, RoutedEventArgs e)
         {
-            // Show AddFreelance window
-            AddFreelance addFreelance = new AddFreelance();
-            addFreelance.Closed += (s, args) => RefreshDataGrid(); // Actualisez le DataGrid après fermeture de la fenêtre d'ajout
-            addFreelance.Show();
+            // Vérifiez si l'utilisateur actuel est administrateur avant d'afficher la fenêtre d'ajout
+            if (Session.CurrentUser != null && Session.CurrentUser.Admin)
+            {
+                AddFreelance addFreelance = new AddFreelance();
+                addFreelance.Closed += (s, args) => RefreshDataGrid();
+                addFreelance.Show();
+            }
+            else
+            {
+                MessageBox.Show("Vous n'êtes pas autorisé à ajouter un nouveau freelance.", "Accès non autorisé", MessageBoxButton.OK, MessageBoxImage.Warning);
+            }
         }
 
-        // Event handler for Remove Button click
         private void RemoveButton_Click(object sender, RoutedEventArgs e)
         {
-            // Check if a freelance is selected
             if (SelectedFreelance != null)
             {
-                using (MakeoProjectContext context = new())
+                if (Session.CurrentUser != null && Session.CurrentUser.Admin)
                 {
-                    // Supprimez d'abord toutes les références au freelance dans la table freelance_projects
-                    var freelanceProjects = context.FreelanceProjects.Where(fp => fp.FreelanceId == SelectedFreelance.Id).ToList();
-                    context.FreelanceProjects.RemoveRange(freelanceProjects);
-                    context.SaveChanges();
+                    MessageBoxResult result = MessageBox.Show("Êtes-vous sûr de vouloir supprimer ce freelance ?", "Confirmation de suppression", MessageBoxButton.YesNo, MessageBoxImage.Warning);
+                    if (result == MessageBoxResult.Yes)
+                    {
+                        using (MakeoProjectContext context = new())
+                        {
+                            var freelanceProjects = context.FreelanceProjects.Where(fp => fp.FreelanceId == SelectedFreelance.Id).ToList();
+                            context.FreelanceProjects.RemoveRange(freelanceProjects);
+                            context.SaveChanges();
 
-                    // Ensuite, supprimez le freelance lui-même
-                    context.Freelances.Remove(SelectedFreelance);
-                    context.SaveChanges();
+                            context.Freelances.Remove(SelectedFreelance);
+                            context.SaveChanges();
 
-                    // Retirez le freelance de la collection allfreelance
-                    allfreelance.Remove(SelectedFreelance);
+                            allfreelance.Remove(SelectedFreelance);
+                            RefreshDataGrid();
 
-                    // Actualisez le DataGrid
-                    RefreshDataGrid();
-
-                    MessageBox.Show("Le freelance a bien été supprimé.", "Confirmation", MessageBoxButton.OK, MessageBoxImage.Information);
+                            MessageBox.Show("Le freelance a bien été supprimé.", "Confirmation", MessageBoxButton.OK, MessageBoxImage.Information);
+                        }
+                    }
+                }
+                else
+                {
+                    MessageBox.Show("Vous n'êtes pas autorisé à supprimer ce freelance.", "Accès non autorisé", MessageBoxButton.OK, MessageBoxImage.Warning);
                 }
             }
             else
@@ -140,16 +138,20 @@ namespace MakeoProject.Views
             }
         }
 
-
-        // Event handler for Edit Button click
         private void EditButton_Click(object sender, RoutedEventArgs e)
         {
-            // Check if a freelance is selected and open the EditFreelance window passing the selected freelance data
             if (SelectedFreelance != null)
             {
-                EditFreelance editFreelance = new EditFreelance(SelectedFreelance);
-                editFreelance.Closed += (s, args) => RefreshDataGrid(); // Actualisez le DataGrid après fermeture de la fenêtre d'édition
-                editFreelance.Show();
+                if (Session.CurrentUser != null && Session.CurrentUser.Id == SelectedFreelance.Id)
+                {
+                    EditFreelance editFreelance = new EditFreelance(SelectedFreelance);
+                    editFreelance.Closed += (s, args) => RefreshDataGrid();
+                    editFreelance.Show();
+                }
+                else
+                {
+                    MessageBox.Show("Vous n'êtes pas autorisé à modifier les données d'un autre freelance.", "Accès non autorisé", MessageBoxButton.OK, MessageBoxImage.Warning);
+                }
             }
             else
             {
@@ -157,7 +159,6 @@ namespace MakeoProject.Views
             }
         }
 
-        // Event handler for Info Button click
         private void InfoButton_Click(object sender, RoutedEventArgs e)
         {
             if (dgFreelances.SelectedItem is Freelance selectedFreelance)
@@ -168,6 +169,14 @@ namespace MakeoProject.Views
             else
             {
                 MessageBox.Show("Veuillez sélectionner un freelance pour voir les détails.", "Avertissement", MessageBoxButton.OK, MessageBoxImage.Warning);
+            }
+        }
+
+        private void dgFreelances_SelectionChanged(object sender, SelectionChangedEventArgs e)
+        {
+            if (dgFreelances.SelectedItem is Freelance selectedFreelance)
+            {
+                SelectedFreelance = selectedFreelance;
             }
         }
     }
